requirements.txt
numpy==1.16.3
easydict==1.9
tqdm==4.32.1
matplotlib==3.1.0
opencv-python==4.1.0.25
Pillow==6.0.0
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
tensorboardX==1.7
torch==1.2.0
torchvision==0.4.0
requests==2.22.0
PyYAML==5.1.2
bcolz==1.2.1
